Rename Timestream Aggregator to Smurf Recorder (#52)
Rename Timestream Aggregator to Smurf Recorder

resolves #49
