+++ agents/smurf_recorder/Dockerfile
@@ -1,14 +1,14 @@
-# SOCS SMuRF Aggregator
+# SOCS SMuRF Recorder
 
 # Use socs base image
 FROM socs:latest
 
 # Set the working directory to registry directory
-WORKDIR /app/socs/agents/timestream_aggregator/
+WORKDIR /app/socs/agents/smurf_recorder/
 COPY . .
 
 # Run registry on container startup
-ENTRYPOINT ["python3", "-u", "timestream_aggregator.py"]
+ENTRYPOINT ["python3", "-u", "smurf_recorder.py"]
 
 # Default site-hub
 CMD ["--site-hub=ws://sisock-crossbar:8001/ws", \
